+++ packages/mobx-keystone-yjs/src/binding/bindYjsToMobxKeystone.ts
@@ -1,3 +1,4 @@
+import { action } from "mobx"
 import {
   AnyDataModel,
   AnyModel,
@@ -13,10 +14,15 @@ import {
   onSnapshot,
 } from "mobx-keystone"
 import * as Y from "yjs"
+import { getOrCreateYjsCollectionAtom } from "../utils/getOrCreateYjsCollectionAtom"
 import { applyMobxKeystonePatchToYjsObject } from "./applyMobxKeystonePatchToYjsObject"
+import { convertYjsDataToJson } from "./convertYjsDataToJson"
 import { convertYjsEventToPatches } from "./convertYjsEventToPatches"
 import { YjsBindingContext, yjsBindingContext } from "./yjsBindingContext"
 
+/**
+ * Creates a bidirectional binding between a Y.js data structure and a mobx-keystone model.
+ */
 export function bindYjsToMobxKeystone<
   TType extends AnyStandardType | ModelClass<AnyModel> | ModelClass<AnyDataModel>,
 >({
@@ -24,15 +30,35 @@ export function bindYjsToMobxKeystone<
   yjsObject,
   mobxKeystoneType,
 }: {
+  /**
+   * The Y.js document.
+   */
   yjsDoc: Y.Doc
-  yjsObject: Y.Map<unknown> | Y.Array<unknown>
+  /**
+   * The bound Y.js data structure.
+   */
+  yjsObject: Y.Map<unknown> | Y.Array<unknown> | Y.Text
+  /**
+   * The mobx-keystone model type.
+   */
   mobxKeystoneType: TType
 }): {
+  /**
+   * The bound mobx-keystone instance.
+   */
   boundObject: TypeToData<TType>
+  /**
+   * Disposes the binding.
+   */
   dispose(): void
+  /**
+   * The Y.js origin symbol used for binding transactions.
+   */
   yjsOrigin: symbol
 } {
   const yjsOrigin = Symbol("bindYjsToMobxKeystoneTransactionOrigin")
+
+  let applyingYjsChangesToMobxKeystone = 0
 
   const bindingContext: YjsBindingContext = {
     yjsDoc,
@@ -40,9 +66,13 @@ export function bindYjsToMobxKeystone<
     mobxKeystoneType,
     yjsOrigin,
     boundObject: undefined, // not yet created
+
+    get isApplyingYjsChangesToMobxKeystone() {
+      return applyingYjsChangesToMobxKeystone > 0
+    },
   }
 
-  const yjsJson = yjsObject.toJSON()
+  const yjsJson = convertYjsDataToJson(yjsObject)
 
   const initializationGlobalPatches: { target: object; patches: Patch[] }[] = []
 
@@ -65,51 +95,55 @@ export function bindYjsToMobxKeystone<
 
   const boundObject = createBoundObject()
 
-  let applyingMobxKeystoneChanges = 0
-
   // bind any changes from yjs to mobx-keystone
-  const observeDeepCb = (events: Y.YEvent<any>[]) => {
+  const observeDeepCb = action((events: Y.YEvent<any>[]) => {
     const patches: Patch[] = []
     events.forEach((event) => {
       if (event.transaction.origin !== yjsOrigin) {
         patches.push(...convertYjsEventToPatches(event))
       }
+
+      if (event.target instanceof Y.Map || event.target instanceof Y.Array) {
+        getOrCreateYjsCollectionAtom(event.target).reportChanged()
+      }
     })
 
     if (patches.length > 0) {
-      applyingMobxKeystoneChanges++
+      applyingYjsChangesToMobxKeystone++
       try {
         applyPatches(boundObject, patches)
       } finally {
-        applyingMobxKeystoneChanges--
+        applyingYjsChangesToMobxKeystone--
       }
     }
-  }
+  })
 
   yjsObject.observeDeep(observeDeepCb)
 
   // bind any changes from mobx-keystone to yjs
-  let pendingPatches: Patch[] = []
+  let pendingArrayOfArrayOfPatches: Patch[][] = []
   const disposeOnPatches = onPatches(boundObject, (patches) => {
-    if (applyingMobxKeystoneChanges > 0) {
+    if (applyingYjsChangesToMobxKeystone > 0) {
       return
     }
 
-    pendingPatches.push(...patches)
+    pendingArrayOfArrayOfPatches.push(patches)
   })
 
   // this is only used so we can transact all patches to the snapshot boundary
   const disposeOnSnapshot = onSnapshot(boundObject, () => {
-    if (pendingPatches.length === 0) {
+    if (pendingArrayOfArrayOfPatches.length === 0) {
       return
     }
 
-    const patches = pendingPatches
-    pendingPatches = []
+    const arrayOfArrayOfPatches = pendingArrayOfArrayOfPatches
+    pendingArrayOfArrayOfPatches = []
 
     yjsDoc.transact(() => {
-      patches.forEach((patch) => {
-        applyMobxKeystonePatchToYjsObject(patch, yjsObject)
+      arrayOfArrayOfPatches.forEach((arrayOfPatches) => {
+        arrayOfPatches.forEach((patch) => {
+          applyMobxKeystonePatchToYjsObject(patch, yjsObject)
+        })
       })
     }, yjsOrigin)
   })