+++ packages/mobx-keystone-yjs/src/index.ts
@@ -1,6 +1,7 @@
+export { YjsTextModel, yjsTextModelId } from "./binding/YjsTextModel"
 export { bindYjsToMobxKeystone } from "./binding/bindYjsToMobxKeystone"
 export {
-  applyJsonArrayYArray,
+  applyJsonArrayToYArray,
   applyJsonObjectToYMap,
   convertJsonToYjsData,
 } from "./binding/convertJsonToYjsData"